requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.14.1
python-dotenv==1.0.1
